Refactor our "message passing"
Switch to index based arguments instead of name based. Hopefully the
performance is better. Also change the way message class is used, so
that it's passed to functions instead of being constructed during
function invocation. It could be in the future also used to store side
effects of function invocation, such as fatal errors.